Revert "[Mbed] Add RPC support for locking and lighting examples (#10960)"
This reverts commit 2149dae4c24add7ad40c8ac89cabb8ce990d78be.
