+++ examples/common/pigweed/rpc_services/Locking.h
@@ -34,14 +34,14 @@ public:
 
     virtual pw::Status Set(ServerContext &, const chip_rpc_LockingState & request, pw_protobuf_Empty & response)
     {
-        bool locked = request.locked;
+        uint8_t locked = request.locked;
         RETURN_STATUS_IF_NOT_OK(app::Clusters::OnOff::Attributes::OnOff::Set(kEndpoint, locked));
         return pw::OkStatus();
     }
 
     virtual pw::Status Get(ServerContext &, const pw_protobuf_Empty & request, chip_rpc_LockingState & response)
     {
-        bool locked;
+        uint8_t locked;
         RETURN_STATUS_IF_NOT_OK(app::Clusters::OnOff::Attributes::OnOff::Get(kEndpoint, &locked));
         response.locked = locked;
         return pw::OkStatus();
