+++ examples/lock-app/mbed/config.in
@@ -3,4 +3,3 @@ CONFIG_CHIP_WITH_EXTERNAL_MBEDTLS=y
 CONFIG_CHIP_PROJECT_CONFIG=main/include/CHIPProjectConfig.h
 CONFIG_CHIP_BYPASS_RENDEZVOUS=n
 CONFIG_MBED_BSD_SOCKET_TRACE=n
-CONFIG_CHIP_PW_RPC=y
