+++ examples/lock-app/mbed/main/AppTask.cpp
@@ -78,6 +78,7 @@ static bool sHaveBLEConnections       = false;
 
 static mbed::Timeout sFunctionTimer;
 
+// TODO: change EventQueue default event size
 static events::EventQueue sAppEventQueue;
 
 using namespace ::chip::Credentials;
@@ -254,31 +255,6 @@ void AppTask::FunctionButtonReleaseEventHandler()
     button_event.ButtonEvent.Pin    = FUNCTION_BUTTON;
     button_event.ButtonEvent.Action = BUTTON_RELEASE_EVENT;
     button_event.Handler            = FunctionHandler;
-    sAppTask.PostEvent(&button_event);
-}
-
-void AppTask::ButtonEventHandler(uint32_t id, bool pushed)
-{
-    if (id > 1)
-    {
-        ChipLogError(NotSpecified, "Wrong button ID");
-        return;
-    }
-
-    AppEvent button_event;
-    button_event.Type               = AppEvent::kEventType_Button;
-    button_event.ButtonEvent.Pin    = id == 0 ? LOCK_BUTTON : FUNCTION_BUTTON;
-    button_event.ButtonEvent.Action = pushed ? BUTTON_PUSH_EVENT : BUTTON_RELEASE_EVENT;
-
-    if (id == 0)
-    {
-        button_event.Handler = LockActionEventHandler;
-    }
-    else
-    {
-        button_event.Handler = FunctionHandler;
-    }
-
     sAppTask.PostEvent(&button_event);
 }
