+++ examples/lock-app/mbed/main/include/AppTask.h
@@ -30,8 +30,6 @@ public:
     void PostEvent(AppEvent * aEvent);
     void UpdateClusterState(void);
 
-    void ButtonEventHandler(uint32_t id, bool pushed);
-
 private:
     friend AppTask & GetAppTask(void);
